+++ src/ixdat/techniques/__init__.py
@@ -8,7 +8,7 @@ Constants:
 """
 
 from .ec import ECMeasurement, ECCalibration
-from .cv import CyclicVoltammogram
+from .cv import CyclicVoltammogram, CyclicVoltammagram  # The latter is deprecated.
 from .ms import MSMeasurement, MSCalibration
 from .ec_ms import ECMSMeasurement, ECMSCalibration
 from .spectroelectrochemistry import SpectroECMeasurement